Minor refactoring in tests
